fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-int fibNumbers[50];
+long fibNumbers[50];
 int i;
 fibNumbers[0] = 1;
 fibNumbers[1] = 2;
